Convert UtilsTest and PresentationTest to StringSpec() + init{}
Matches the style used by the new ConfigsTest, OutputConfigTest,
LetsPlotTest, and LetsPlotDslTest suites. Also picks up the test
file moves into com/kslides/ package directories done by the IDE.
